node: move archive gql code into own module
